backend/requirements.txt
fastapi
uvicorn
firebase-admin
tensorflow
xgboost
scikit-learn
pandas
numpy
python-multipart
joblib
python-dotenv
